pyproject.toml
[project]
name = "django_template"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "a-paudel", email = "ankur@paudel.me"},
]
dependencies = [
    "django>=5.0.1",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "dj-database-url>=2.1.0",
    "gunicorn>=21.2.0",
    "psycopg[binary]>=3.1.17",
    "whitenoise[brotli]>=6.6.0",
    "django-debug-toolbar>=4.2.0",
    "django-unfold>=0.20.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
